fix: export and use the correct `UmbCropModel` everywhere along with the rest of the types
