+++ src/packages/media/media/components/input-image-cropper/types.ts
@@ -1,17 +1,9 @@
+import type { UmbCropModel, UmbFocalPointModel } from '../../types.js';
+
 export type UmbImageCropperPropertyEditorValue = {
 	temporaryFileId?: string | null;
-	crops: Array<{
-		alias: string;
-		coordinates?: {
-			x1: number;
-			x2: number;
-			y1: number;
-			y2: number;
-		};
-		height: number;
-		width: number;
-	}>;
-	focalPoint: { left: number; top: number };
+	crops: Array<UmbCropModel>;
+	focalPoint: UmbFocalPointModel;
 	src: string;
 };
